Applied requested changes for the network command
