+++ src/o365/base/YammerCommand.ts
@@ -1,7 +1,16 @@
-import Command from "../../Command";
+import Command, { CommandError } from "../../Command";
 
-export default abstract class GraphCommand extends Command {
+export default abstract class YammerCommand extends Command {
   protected get resource(): string {
     return 'https://www.yammer.com/api';
+  }
+
+  protected handleRejectedODataJsonPromise(response: any, cmd: CommandInstance, callback: (err?: any) => void): void {
+    if (response.error && response.error.base) {
+      callback(new CommandError(response.error.base));
+    }
+    else {
+      callback(new CommandError(response));
+    }
   }
 }
